chanstate: move db codec
Move the shared database element codec into the chanstate package so
channel-state serialization helpers can use it without depending on
channeldb. Leave compatibility wrappers in channeldb for existing
callers and keep the peer flap time serialization helpers there because
they are not channel-state code.
